whatWeDo folder's name under public folder and component folder are changed. Folders are organised  and imports updated accordingly
